requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-dotenv==1.0.1
faiss-cpu==1.7.4
numpy==1.26.4
openai==1.12.0
anthropic==0.18.1
sentence-transformers==2.5.1
pydantic==2.6.1
python-multipart==0.0.9 